+++ src/collision/shapes/ConvexMeshShape.cpp
@@ -39,8 +39,8 @@ using namespace reactphysics3d;
  * @param stride Stride between the beginning of two elements in the vertices array
  * @param margin Collision margin (in meters) around the collision shape
  */
-ConvexMeshShape::ConvexMeshShape(PolyhedronMesh* polyhedronMesh, MemoryAllocator& allocator, const Vector3& scale)
-                : ConvexPolyhedronShape(CollisionShapeName::CONVEX_MESH, allocator), mPolyhedronMesh(polyhedronMesh),
+ConvexMeshShape::ConvexMeshShape(ConvexMesh* convexMesh, MemoryAllocator& allocator, const Vector3& scale)
+                : ConvexPolyhedronShape(CollisionShapeName::CONVEX_MESH, allocator), mConvexMesh(convexMesh),
                   mMinBounds(0, 0, 0), mMaxBounds(0, 0, 0), mScale(scale) {
 
     // Recalculate the bounds of the mesh
@@ -61,10 +61,10 @@ Vector3 ConvexMeshShape::getLocalSupportPointWithoutMargin(const Vector3& direct
     uint32 indexMaxDotProduct = 0;
 
     // For each vertex of the mesh
-    for (uint32 i=0; i<mPolyhedronMesh->getNbVertices(); i++) {
+    for (uint32 i=0; i<mConvexMesh->getNbVertices(); i++) {
 
         // Compute the dot product of the current vertex
-        decimal dotProduct = direction.dot(mPolyhedronMesh->getVertex(i));
+        decimal dotProduct = direction.dot(mConvexMesh->getVertex(i));
 
         // If the current dot product is larger than the maximum one
         if (dotProduct > maxDotProduct) {
@@ -76,26 +76,26 @@ Vector3 ConvexMeshShape::getLocalSupportPointWithoutMargin(const Vector3& direct
     assert(maxDotProduct >= decimal(0.0));
 
     // Return the vertex with the largest dot product in the support direction
-    return mPolyhedronMesh->getVertex(indexMaxDotProduct) * mScale;
+    return mConvexMesh->getVertex(indexMaxDotProduct) * mScale;
 }
 
 // Recompute the bounds of the mesh
 void ConvexMeshShape::recalculateBounds() {
 
-    mMinBounds = mPolyhedronMesh->getVertex(0);
-    mMaxBounds = mPolyhedronMesh->getVertex(0);
+    mMinBounds = mConvexMesh->getVertex(0);
+    mMaxBounds = mConvexMesh->getVertex(0);
 
     // For each vertex of the mesh
-    for (uint32 i=1; i<mPolyhedronMesh->getNbVertices(); i++) {
+    for (uint32 i=1; i<mConvexMesh->getNbVertices(); i++) {
 
-        if (mPolyhedronMesh->getVertex(i).x > mMaxBounds.x) mMaxBounds.x = mPolyhedronMesh->getVertex(i).x;
-        if (mPolyhedronMesh->getVertex(i).x < mMinBounds.x) mMinBounds.x = mPolyhedronMesh->getVertex(i).x;
+        if (mConvexMesh->getVertex(i).x > mMaxBounds.x) mMaxBounds.x = mConvexMesh->getVertex(i).x;
+        if (mConvexMesh->getVertex(i).x < mMinBounds.x) mMinBounds.x = mConvexMesh->getVertex(i).x;
 
-        if (mPolyhedronMesh->getVertex(i).y > mMaxBounds.y) mMaxBounds.y = mPolyhedronMesh->getVertex(i).y;
-        if (mPolyhedronMesh->getVertex(i).y < mMinBounds.y) mMinBounds.y = mPolyhedronMesh->getVertex(i).y;
+        if (mConvexMesh->getVertex(i).y > mMaxBounds.y) mMaxBounds.y = mConvexMesh->getVertex(i).y;
+        if (mConvexMesh->getVertex(i).y < mMinBounds.y) mMinBounds.y = mConvexMesh->getVertex(i).y;
 
-        if (mPolyhedronMesh->getVertex(i).z > mMaxBounds.z) mMaxBounds.z = mPolyhedronMesh->getVertex(i).z;
-        if (mPolyhedronMesh->getVertex(i).z < mMinBounds.z) mMinBounds.z = mPolyhedronMesh->getVertex(i).z;
+        if (mConvexMesh->getVertex(i).z > mMaxBounds.z) mMaxBounds.z = mConvexMesh->getVertex(i).z;
+        if (mConvexMesh->getVertex(i).z < mMinBounds.z) mMinBounds.z = mConvexMesh->getVertex(i).z;
     }
 
     // Apply the local scaling factor
@@ -116,15 +116,15 @@ bool ConvexMeshShape::raycast(const Ray& ray, RaycastInfo& raycastInfo, Collider
     Vector3 currentFaceNormal;
     bool isIntersectionFound = false;
 
-    const HalfEdgeStructure& halfEdgeStructure = mPolyhedronMesh->getHalfEdgeStructure();
+    const HalfEdgeStructure& halfEdgeStructure = mConvexMesh->getHalfEdgeStructure();
 
     // For each face of the convex mesh
-    for (uint32 f=0; f < mPolyhedronMesh->getNbFaces(); f++) {
+    for (uint32 f=0; f < mConvexMesh->getNbFaces(); f++) {
 
         const HalfEdgeStructure::Face& face = halfEdgeStructure.getFace(f);
-        const Vector3& faceNormal = mPolyhedronMesh->getFaceNormal(f);
+        const Vector3& faceNormal = mConvexMesh->getFaceNormal(f);
         const HalfEdgeStructure::Vertex& faceVertex = halfEdgeStructure.getVertex(face.faceVertices[0]);
-        const Vector3& facePoint = mPolyhedronMesh->getVertex(faceVertex.vertexPointIndex);
+        const Vector3& facePoint = mConvexMesh->getVertex(faceVertex.vertexPointIndex);
         decimal denom = faceNormal.dot(direction);
         decimal planeD = faceNormal.dot(facePoint);
         decimal dist = planeD -  faceNormal.dot(ray.point1);
@@ -186,15 +186,15 @@ bool ConvexMeshShape::raycast(const Ray& ray, RaycastInfo& raycastInfo, Collider
 // Return true if a point is inside the collision shape
 bool ConvexMeshShape::testPointInside(const Vector3& localPoint, Collider* /*collider*/) const {
 
-    const HalfEdgeStructure& halfEdgeStructure = mPolyhedronMesh->getHalfEdgeStructure();
+    const HalfEdgeStructure& halfEdgeStructure = mConvexMesh->getHalfEdgeStructure();
 
     // For each face plane of the convex mesh
-    for (uint32 f=0; f < mPolyhedronMesh->getNbFaces(); f++) {
+    for (uint32 f=0; f < mConvexMesh->getNbFaces(); f++) {
 
         const HalfEdgeStructure::Face& face = halfEdgeStructure.getFace(f);
-        const Vector3& faceNormal = mPolyhedronMesh->getFaceNormal(f);
+        const Vector3& faceNormal = mConvexMesh->getFaceNormal(f);
         const HalfEdgeStructure::Vertex& faceVertex = halfEdgeStructure.getVertex(face.faceVertices[0]);
-        const Vector3& facePoint = mPolyhedronMesh->getVertex(faceVertex.vertexPointIndex);
+        const Vector3& facePoint = mConvexMesh->getVertex(faceVertex.vertexPointIndex);
 
         // If the point is out of the face plane, it is outside of the convex mesh
         if (computePointToPlaneDistance(localPoint, faceNormal, facePoint) > decimal(0.0)) return false;
@@ -208,24 +208,24 @@ std::string ConvexMeshShape::to_string() const {
 
     std::stringstream ss;
     ss << "ConvexMeshShape{" << std::endl;
-    ss << "nbVertices=" << mPolyhedronMesh->getNbVertices() << std::endl;
-    ss << "nbFaces=" << mPolyhedronMesh->getNbFaces() << std::endl;
+    ss << "nbVertices=" << mConvexMesh->getNbVertices() << std::endl;
+    ss << "nbFaces=" << mConvexMesh->getNbFaces() << std::endl;
 
     ss << "vertices=[";
 
-    for (uint32 v=0; v < mPolyhedronMesh->getNbVertices(); v++) {
+    for (uint32 v=0; v < mConvexMesh->getNbVertices(); v++) {
 
-        const Vector3& vertex = mPolyhedronMesh->getVertex(v);
+        const Vector3& vertex = mConvexMesh->getVertex(v);
         ss << vertex.to_string();
-        if (v != mPolyhedronMesh->getNbVertices() - 1) {
+        if (v != mConvexMesh->getNbVertices() - 1) {
             ss << ", ";
         }
     }
 
     ss << "], faces=[";
 
-    HalfEdgeStructure halfEdgeStruct = mPolyhedronMesh->getHalfEdgeStructure();
-    for (uint32 f=0; f < mPolyhedronMesh->getNbFaces(); f++) {
+    HalfEdgeStructure halfEdgeStruct = mConvexMesh->getHalfEdgeStructure();
+    for (uint32 f=0; f < mConvexMesh->getNbFaces(); f++) {
 
         const HalfEdgeStructure::Face& face = halfEdgeStruct.getFace(f);